Hoist shared streaming I/O loop and unpack helpers into stream_io (Issue #203)
The head-and-compact .bin streaming loop and its byte-level helpers
(unpack_f32s_le, reserve_unpack_capacity, compact_pending_if_needed) were
duplicated between multi_score.rs and stream_score.rs, leaving the Issue #103
OOB-safety invariant in two copies to keep in lock-step.

Hoist them plus run_io_loop / ScoreChunkFn into a new shared crate::stream_io
module. The fused forward-only path (stream_score) and the CPU directory path
(multi_score) now feed a score_chunk closure into the shared run_io_loop; the
GPU directory path already did. One copy of each helper, one loop.

The relocated unpack_f32s_le #103 should_panic tests now live (once) beside
the function in stream_io; the duplicate copies in the two modules were
removed. New unit tests cover run_io_loop fast/slow paths and error
propagation.

Closes #203.

Vibe-Coder-Run-Id: vibe-mqbytmhz-848e5b
